+++ src/modules/linkdrop-sdk/index.ts
@@ -1,6 +1,16 @@
-import { ILinkdropSDK, TNetworkName } from '../../types'
+import {
+  ILinkdropSDK,
+  TNetworkName,
+  TMode
+} from '../../types'
 import Campaign from '../campaign'
-import { defineCampaignSig, getLinkParams, getLinkStatus, redeemLink } from '../../helpers'
+import {
+  defineCampaignSig,
+  getLinkParams,
+  getLinkStatus,
+  redeemLink,
+  defineClaimHostUrl
+} from '../../helpers'
 import { campaignsApi } from '../../api'
 import {
   testnetsApiUrl,
@@ -15,6 +25,8 @@ class LinkdropSDK implements ILinkdropSDK {
   apiHost: string
   claimHostUrl: string
   apiKey: string
+  mode: TMode
+
   utils = {
     createLink,
     computeProxyAddress,
@@ -29,10 +41,13 @@ class LinkdropSDK implements ILinkdropSDK {
   }: {
     apiKey?: string,
     apiHost?: string,
-    mode?: 'testnets',
+    mode?: TMode,
     claimHostUrl?: string
   } = {}) {
     this.claimHostUrl = claimHostUrl || ''
+
+    this.mode = mode || 'mainnets'
+
     if (!apiKey) {
       throw new Error('ApiKey required')
     } else {
@@ -41,7 +56,7 @@ class LinkdropSDK implements ILinkdropSDK {
     if (apiHost) {
       this.apiHost = apiHost
     } else {
-      this.apiHost = mode === 'testnets' ? testnetsApiUrl : apiUrl
+      this.apiHost = this.mode === 'testnets' ? testnetsApiUrl : apiUrl
     }
   }
 
@@ -78,7 +93,7 @@ class LinkdropSDK implements ILinkdropSDK {
         signerKey,
         encryptionKey,
         campaign,
-        this.claimHostUrl,
+        this.claimHostUrl ? this.claimHostUrl : defineClaimHostUrl(this.mode),
         campaignSig,
         this.apiHost,
         this.apiKey