+++ src/types/mode/index.ts
@@ -1,0 +1,3 @@
+type TMode = 'testnets' | 'mainnets'
+
+export default TMode